Websocket Proxy now enabled
This is a first iteration of the webproxy and standard tcp client proxy.
I used novnc as the viewer and Ultravnc as the server and everything
worked properly. BE AWARE, THIS IS HIGHLY UNOPTIMIZED! I am releasing
this now just as a proof of concept and will release another update
cleaning everything up.   This utilizes an amazing websocket server that
you can check out here https://github.com/rickparrish/Fleck